+++ src/theo_comm/include/broker_codes.hpp
@@ -1,13 +1,15 @@
 #pragma once
 
+#include <cstdint>
+
 // Broker State Machine Codes
 // 0 |-> confirm to next state
 // 1 |-> reset
 // 2 |-> check broker state
 
-enum class BroadcastBrokerCodes : int {
+enum class BroadcastBrokerCodes : int8_t {
     Confirm,// confirm to next state
-    Reset,  // received broadcast request from transmitter
-    Check,  // robot confirmed broadcast request, transmitter may broadcast to shared topic
+    Reset,  // reset brokerage to IDLE state
+    Check,  // request information on broker state
 };
